Implemating route and minor style
