+++ src/app/app.component.html
@@ -1,36 +1,1 @@
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
-<link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-
-<div class="col-1" style="text-align: center">
-    <div class="searchBoxDataset">
-      <span>Enter your search : </span>
-      <input id="input-search" type="search" placeholder="Search" [(ngModel)]='searchStr'>
-      <button type="button" class="link">
-        <span class="material-icons" style="font-size: 15px" (click)="onSearch(searchStr)">search</span>
-      </button>
-      <div *ngIf="errorMsg" class="text-danger">
-        <div>{{errorMsg}}</div>
-      </div>
-    </div>
-</div>
-
-<div class="col-2">
-    <label class="lable"> Group By </label>
-    <select class="select " [(ngModel)]="selectedGroupBy" (change)='resetDataCardView()'
-      name="groupByOptions" id="data_groupby" class="dropdown" required=true>
-      <option *ngFor='let option of groupByOptions' [ngValue]="option">
-        {{option.key}}</option>
-    </select>
-</div>
-<div class="col-3">
-  <mat-grid-list cols="3" rowHight="2:1" gutterSize="10px">
-    <div *ngFor='let data of cardDataToDisplay'>
-      <mat-grid-tile>
-        <div>Type: {{ data.type }}</div>
-        <div> Name: {{ data.name }}</div>
-
-      </mat-grid-tile>
-    </div>
-  </mat-grid-list>
-</div>
+<router-outlet></router-outlet>
